+++ paddle/pten/api/lib/kernel_declare.h
@@ -26,6 +26,7 @@ PT_DECLARE_KERNEL(flatten, CPU, ALL_LAYOUT);
 PT_DECLARE_KERNEL(sign, CPU, ALL_LAYOUT);
 PT_DECLARE_KERNEL(scale, CPU, ALL_LAYOUT);
 PT_DECLARE_KERNEL(conj, CPU, ALL_LAYOUT);
+PT_DECLARE_KERNEL(add, CPU, ALL_LAYOUT);
 
 #if defined(PADDLE_WITH_CUDA) || defined(PADDLE_WITH_HIP)
 PT_DECLARE_KERNEL(full_like, CUDA, ALL_LAYOUT);
@@ -34,6 +35,7 @@ PT_DECLARE_KERNEL(flatten, CUDA, ALL_LAYOUT);
 PT_DECLARE_KERNEL(sign, CUDA, ALL_LAYOUT);
 PT_DECLARE_KERNEL(scale, CUDA, ALL_LAYOUT);
 PT_DECLARE_KERNEL(conj, CUDA, ALL_LAYOUT);
+PT_DECLARE_KERNEL(add, CUDA, ALL_LAYOUT);
 #endif
 
 #ifdef PADDLE_WITH_XPU
