+++ paddle/pten/kernels/cuda/CMakeLists.txt
@@ -6,6 +6,7 @@ if(WITH_GPU)
   nv_library(scale_kernel_cuda SRCS scale_kernel.cu DEPS dense_tensor kernel_context kernel_factory eigen_function)
   nv_library(full_kernel_cuda SRCS full_kernel.cu DEPS dense_tensor kernel_context kernel_factory eigen_function)
   nv_library(conj_kernel_cuda SRCS conj_kernel.cu DEPS dense_tensor kernel_context kernel_factory)
+  nv_library(elementwise_kernel_cuda SRCS elementwise_kernel.cu DEPS dense_tensor convert_utils kernel_context kernel_factory)
 elseif(WITH_ROCM)
   hip_library(math_cuda SRCS math.cu DEPS eigen_function dense_tensor convert_utils kernel_context kernel_factory pten_transpose_cuda)
   hip_library(linalg_cuda SRCS linalg.cu DEPS eigen_function dense_tensor kernel_context kernel_factory)
@@ -14,4 +15,5 @@ elseif(WITH_ROCM)
   hip_library(scale_kernel_cuda SRCS scale_kernel.cu DEPS dense_tensor kernel_context kernel_factory eigen_function)
   hip_library(full_kernel_cuda SRCS full_kernel.cu DEPS dense_tensor kernel_context kernel_factory eigen_function)
   hip_library(conj_kernel_cuda SRCS conj_kernel.cu DEPS dense_tensor kernel_context kernel_factory)
+  hip_library(elementwise_kernel_cuda SRCS elementwise_kernel.cu DEPS dense_tensor convert_utils kernel_context kernel_factory)
 endif()
